chore: create IndividualBar class
This commit adds the `IndividualBar` class, which represents an individual bar
in the bar graph. The class includes properties `x` and `y` to define the position
and height of the bar, respectively.

This class is intended to be used in the context of the `MyBarGraph` widget to
facilitate the representation of data for each bar in the bar chart.